Move Style Book frontend assets from includes/ to src/.
Align the Style Book bundle with the project convention of keeping admin JS under src/ while PHP stays in includes/resources/.
