+++ src/style-book.js
@@ -2,8 +2,8 @@
  * Style Book admin entry.
  */
 import { createRoot } from '@wordpress/element';
-import { StyleBookApp } from './app/StyleBookApp';
-import './style-book.scss';
+import { StyleBookApp } from './style-book/app/StyleBookApp';
+import './style-book/style-book.scss';
 
 wp.domReady( () => {
 	const container = document.getElementById( 'kadence-style-book-root' );
